Performance Enhancements:
Removed unnecessary default field initializations:  Every field is assigned the value zero or null when a class is instantiated by the virtual machine. Explicitly
assigning these values is unnecessary, and can make debugging more painful, because the assignments count as a statement.

Used Boolean.valueOf vs. new Boolean().

Removed unnecessary Cast, as some compilers do not optimize.

Changed '|' to '||' in 'if' statement, as seemed to be a typographical error.

Added ';' and comments to an empty block.


SVN-Revision: 126